Parse Sink Messages inside the client and return human-readable content.
For Netflow data, it returns a JSON representation.
For Syslog data, it returns a JSON representation of the decoded XML.
For Flow Document, or enriched content used by Nephron, similar to
Netflow, it returns a JSON representation.
Logs were improved, as well as the CLI and Docker entrypoint.

Pending: SNMP for Traps